outbox/requirements.txt
psycopg2-binary==3.3.2

